# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: SGMA APP2 API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /admin/users:
    get:
      operationId: getAdminUsers
      tags: [admin]
      summary: List users (admin/super only)
      security:
        - BearerAuth: []
      parameters:
        - name: q
          in: query
          required: false
          schema:
            type: string
        - name: role
          in: query
          required: false
          schema:
            type: string
            enum: [MEMBER, MODERATOR, ADMIN, SUPER_ADMIN]
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum: [PENDING, ACTIVE, SUSPENDED]
      responses:
        "200":
          description: Users list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUsersListResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/users/{id}:
    get:
      operationId: getAdminUser
      tags: [admin]
      summary: Get one user (admin/super only)
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUserDetail"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateAdminUser
      tags: [admin]
      summary: Update a user's profile/role/status (admin/super only, permission-checked)
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminUpdateUserInput"
      responses:
        "200":
          description: User updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUserDetail"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    AdminUserItem:
      type: object
      required: [id, fullName, account, email, role, status, isActive, createdAt]
      properties:
        id:
          type: integer
        fullName:
          type: string
        account:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [MEMBER, MODERATOR, ADMIN, SUPER_ADMIN]
        status:
          type: string
          enum: [PENDING, ACTIVE, SUSPENDED]
        isActive:
          type: boolean
        professionGroup:
          type: ["string", "null"]
        membershipNumber:
          type: ["string", "null"]
        accessScope:
          type: string
          enum: [FULL_APP, SYRIA_ACADEMY_ONLY]
        academySpecialty:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
    AdminUsersListResponse:
      type: array
      items:
        $ref: "#/components/schemas/AdminUserItem"
    AdminUserDetail:
      type: object
      required: [id, fullName, account, email, role, status, isActive, isDeveloper, createdAt, updatedAt]
      properties:
        id:
          type: integer
        fullName:
          type: string
        account:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [MEMBER, MODERATOR, ADMIN, SUPER_ADMIN]
        status:
          type: string
          enum: [PENDING, ACTIVE, SUSPENDED]
        isActive:
          type: boolean
        isDeveloper:
          type: boolean
        phone:
          type: ["string", "null"]
        whatsapp:
          type: ["string", "null"]
        birthDate:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        professionGroup:
          type: ["string", "null"]
        specialtyText:
          type: ["string", "null"]
        bio:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]
        membershipNumber:
          type: ["string", "null"]
        accessScope:
          type: string
          enum: [FULL_APP, SYRIA_ACADEMY_ONLY]
        academySpecialty:
          type: ["string", "null"]
        academyAllowedSpecialties:
          type: array
          items:
            type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    AdminUpdateUserInput:
      type: object
      properties:
        fullName:
          type: string
          minLength: 1
        account:
          type: string
          minLength: 1
        email:
          type: string
        phone:
          type: ["string", "null"]
        whatsapp:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        birthDate:
          type: ["string", "null"]
        professionGroup:
          type: ["string", "null"]
        specialtyText:
          type: ["string", "null"]
        bio:
          type: ["string", "null"]
        membershipNumber:
          type: ["string", "null"]
        role:
          type: string
          enum: [MEMBER, MODERATOR, ADMIN, SUPER_ADMIN]
        status:
          type: string
          enum: [PENDING, ACTIVE, SUSPENDED]
        isActive:
          type: boolean
        accessScope:
          type: string
          enum: [FULL_APP, SYRIA_ACADEMY_ONLY]
        academySpecialty:
          type: ["string", "null"]
        academyAllowedSpecialties:
          type: array
          items:
            type: string
...